Cambios en la creación de cursos

+++ src/controllers/CursoController.py
@@ -1,6 +1,6 @@
 import math
 from flask import render_template, jsonify, request, flash, redirect, url_for
-from forms import FormCurso, UpdateYearCursoForm
+from forms import FormCurso, FormCursoUpdate, UpdateYearCursoForm
 from models.Asignatura import Asignatura
 from models.Curso import Curso
 from models.CursoAsignatura import CursoAsignatura, Modalidad
@@ -16,43 +16,53 @@ def index():
 
 def add():
     formulario = FormCurso()
-    # asignaturas = Asignatura.get_asignaturas_groupby_titulacion()
-    # asignaturas = Asignatura.get_all_json()
     if formulario.validate_on_submit():
         ano_inicio = formulario.ano_inicio.data
         curso = Curso(ano_inicio=ano_inicio)
         db.session.add(curso)
-        db.session.flush()
+        db.session.commit()
+        flash('Curso creado correctamente', 'alert alert-success alert-dismissible fade show')
+        return redirect(url_for('curso_bp.update', id_curso=curso.id))
+    return render_template('cursos/form-create.html', form=formulario)
 
-        # Curso_Asignatura
+
+def update(id_curso):
+    curso = Curso.get_curso(id_curso)
+    formulario = FormCursoUpdate(obj=curso)
+    list_asignaturas = list(set([str(curso_asignatura.id_asignatura) for curso_asignatura in curso.asignaturas]))
+    # asignaturas_actuales = list(set([curso_asignatura.asignatura for curso_asignatura in curso.asignaturas]))
+
+    if curso is None:
+        flash('Curso no encontrado', 'alert alert-danger alert-dismissible fade show')
+        return redirect(url_for('curso_bp.index'))
+    if formulario.validate_on_submit():
         alumnos_presencial = formulario.n_a_p.data
         alumnos_ingles = formulario.n_a_i.data
         alumnos_online = formulario.n_a_o.data
 
-        for id_asignatura in formulario.id_asignaturas.data.split(','):
-            if alumnos_presencial > 0:
-                num_grupos_teoricos = formulario.n_g_t_p.data
-                num_grupos_practicos = formulario.n_g_p_p.data
-                create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id, alumnos_presencial,
-                                Modalidad.Presencial)
+        asignaturas_nuevas = formulario.id_asignaturas.data.split(',')
+        asignaturas_to_add = list(set(asignaturas_nuevas) - set(list_asignaturas))
 
-            if alumnos_ingles > 0:
-                num_grupos_teoricos = formulario.n_g_t_i.data
-                num_grupos_practicos = formulario.n_g_p_i.data
-                create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id, alumnos_ingles,
-                                Modalidad.Ingles)
+        if len(asignaturas_to_add) > 0:
+            for id_asignatura in asignaturas_to_add:
+                if alumnos_presencial > 0:
+                    create_relation(1, 1, id_asignatura, curso.id, alumnos_presencial, Modalidad.Presencial)
 
-            if alumnos_online > 0:
-                num_grupos_teoricos = formulario.n_g_t_o.data
-                num_grupos_practicos = formulario.n_g_p_o.data
-                create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id, alumnos_online,
-                                Modalidad.Online)
+                if alumnos_ingles > 0:
+                    create_relation(1, 1, id_asignatura, curso.id, alumnos_ingles, Modalidad.Ingles)
+
+                if alumnos_online > 0:
+                    create_relation(1, 1, id_asignatura, curso.id, alumnos_online, Modalidad.Online)
 
         db.session.commit()
-        flash('Curso creado correctamente', 'alert alert-success alert-dismissible fade show')
 
+        flash('Asignaturas y grupos añadidos correctamente', 'alert alert-success alert-dismissible fade show')
         return redirect(url_for('curso_bp.index'))
-    return render_template('cursos/form.html', form=formulario, asignaturas=[])
+
+    formulario.id_asignaturas.data = ','.join(
+        set([str(curso_asignatura.id_asignatura) for curso_asignatura in curso.asignaturas]))
+    return render_template('cursos/form-update.html', form=formulario, asig_actuales=[],
+                           id_curso=id_curso)
 
 
 def create_relation(n_g_t, n_g_p, id_asignatura, id_curso, alumnos, modalidad):
@@ -105,7 +115,8 @@ def render_sortable_edit():
         asignaturas_curso = []
         for curso_asignatura in cursos_asignaturas:
             asignaturas_curso.append(curso_asignatura.asignatura)
-        asignaturas = Asignatura.get_asignaturas_by_titulacion(request.form.get('id_titulacion'))
+        asignaturas = Asignatura.get_asignaturas_by_titulacion_curso(request.form.get('id_titulacion'),
+                                                                     request.form.get('curso'))
 
         return jsonify(
             render_template('cursos/sortable.html', asignaturas=list(set(asignaturas) - set(asignaturas_curso))))
@@ -114,6 +125,10 @@ def render_sortable_edit():
 def delete(id_curso):
     curso = Curso.get_curso(id_curso)
     if curso is not None:
+        if curso.asignaturas:
+            flash('No se puede eliminar el curso porque tiene asignaturas asociadas',
+                  'alert alert-danger alert-dismissible fade show')
+            return redirect(url_for('curso_bp.index'))
         db.session.delete(curso)
         db.session.commit()
         flash('Curso eliminado correctamente', 'alert alert-success alert-dismissible fade show')
@@ -174,61 +189,3 @@ def duplicate(id_curso):
         flash('Error al duplicar el curso: ' + str(e), 'alert alert-danger alert-dismissible fade show')
 
     return redirect(url_for('curso_bp.index'))
-
-
-def update(id_curso):
-    curso = Curso.get_curso(id_curso)
-    formulario = FormCurso(obj=curso)
-    list_asignaturas = list(set([str(curso_asignatura.id_asignatura) for curso_asignatura in curso.asignaturas]))
-    asignaturas_actuales = list(set([curso_asignatura.asignatura for curso_asignatura in curso.asignaturas]))
-    formulario.submit.label.text = 'Modificar'
-
-    if curso is None:
-        flash('Curso no encontrado', 'alert alert-danger alert-dismissible fade show')
-        return redirect(url_for('curso_bp.index'))
-    if formulario.validate_on_submit():
-        curso.ano_inicio = formulario.ano_inicio.data
-
-        alumnos_presencial = formulario.n_a_p.data
-        alumnos_ingles = formulario.n_a_i.data
-        alumnos_online = formulario.n_a_o.data
-
-        asignaturas_nuevas = formulario.id_asignaturas.data.split(',')
-        asignaturas_to_add = list(set(asignaturas_nuevas) - set(list_asignaturas))
-
-        if len(asignaturas_to_add) > 0:
-            for id_asignatura in asignaturas_to_add:
-                if alumnos_presencial > 0:
-                    num_grupos_teoricos = formulario.n_g_t_p.data
-                    num_grupos_practicos = formulario.n_g_p_p.data
-                    create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id,
-                                    alumnos_presencial,
-                                    Modalidad.Presencial)
-
-                if alumnos_ingles > 0:
-                    num_grupos_teoricos = formulario.n_g_t_i.data
-                    num_grupos_practicos = formulario.n_g_p_i.data
-                    create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id, alumnos_ingles,
-                                    Modalidad.Ingles)
-
-                if alumnos_online > 0:
-                    num_grupos_teoricos = formulario.n_g_t_o.data
-                    num_grupos_practicos = formulario.n_g_p_o.data
-                    create_relation(num_grupos_teoricos, num_grupos_practicos, id_asignatura, curso.id, alumnos_online,
-                                    Modalidad.Online)
-
-        delete_elements = list(set(list_asignaturas) - set(asignaturas_nuevas))
-        if len(delete_elements) > 0:
-            for id_asignatura in delete_elements:
-                cursos_asignaturas = CursoAsignatura.get_curso_asignatura(id_asignatura, curso.id)
-                if cursos_asignaturas is not None:
-                    for curso_asignatura in cursos_asignaturas:
-                        db.session.delete(curso_asignatura)
-        db.session.commit()
-
-        flash('Curso actualizado correctamente', 'alert alert-success alert-dismissible fade show')
-        return redirect(url_for('curso_bp.index'))
-    formulario.id_asignaturas.data = ','.join(
-        set([str(curso_asignatura.id_asignatura) for curso_asignatura in curso.asignaturas]))
-    return render_template('cursos/form-update.html', form=formulario, asig_actuales=asignaturas_actuales,
-                           id_curso=id_curso)
